api/requirements.txt
fastapi 
uvicorn 
chromadb
sentence-transformers 
crawl4ai
pydantic 
python-dotenv 
requests